% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia.R
\name{julia}
\alias{julia}
\title{Generate a julia set}
\usage{
julia(
  cx = -0.7,
  cy = 0.27015,
  movex = 0,
  movey = 0,
  zoom = 1,
  size = 400,
  max_iter = 255,
  gamma = 1,
  equalize = FALSE
)
}
\arguments{
\item{cx, cy}{real and imaginary components of 'c'}

\item{movex, movey}{offsets in centre position}

\item{zoom}{zoom level. default 1.}

\item{size}{width and height of resulting array}

\item{max_iter}{maximum number of iterations}

\item{gamma}{Gamma correction exponent. Default: 1}

\item{equalize}{Equalize histogram to enhance the contrast. default: FALSE}
}
\value{
raw vector with dimensions (size, size)
}
\description{
This function generates Julia fractals as an array of raw values.  The
values represent the number of iterations performed before the values at
the pixel location iterated towards infinity.  It is not a raw count of iterations,
but rather the count of iterations scaled to be a maximum of 255 - and thus
the data fits in a \code{raw} vector.
}
\examples{
\dontrun{
js <- julia()
mode(js) <- 'integer'
plot(as.raster(js/255))
}
}
