% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia-base.R
\name{julia_r}
\alias{julia_r}
\title{Calculate Julia set using base R matrix operations}
\usage{
julia_r(c_re = -0.7, c_im = 0.27015, width = 400, height = 200, max_iter = 255)
}
\arguments{
\item{c_re, c_im}{real and imaginary components of 'c'. Default [-0.7, 0.27015]}

\item{width, height}{size of the output image default: 400x400}

\item{max_iter}{maximum number of iterations. Default: 255}
}
\value{
matrix of iteration count
}
\description{
Users are encouraged to copy the body of this function and hack on it in 
order to achieve novel renderings.
}
\examples{
# Calculate number of iterations for each location with default parameters
iters <- julia_r()

# Specify the colour scheme
cols <- terrain.colors(256) 

# Convert the integer iteration count to the range [1, 256]
# Do a little bit of gamma correction while we're at it
z <- round( (iters / max(iters)) ^ 0.25 * 255) + 1

# Convert from integer to colour
z[] <- cols[z]
plot(as.raster(z), interpolate = FALSE)
}
