% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia-base.R
\name{julia_r}
\alias{julia_r}
\title{Calculate Julia set using base R matrix operations}
\usage{
julia_r(cx = -0.7, cy = 0.27015, size = 400, max_iter = 255, gamma = 1)
}
\arguments{
\item{cx, cy}{real and imaginary components of 'c'. Default [-0.7, 0.27015]}

\item{size}{width and height of resulting square array. Default: 400}

\item{max_iter}{maximum number of iterations. Default: 256}

\item{gamma}{Gamma correction exponent. Default: 1}
}
\value{
matrix of iteration count
}
\description{
Calculate Julia set using base R matrix operations
}
\examples{
iters <- julia_r()
}
